chore: move winnerTable component to new folder
