+++ src/pages/winners/ui/winners-page.ts
@@ -1,7 +1,7 @@
 import { winnersStore, winnersQueryStore, WINNERS_ON_PAGE } from '@/entities';
 import { BlockComponent } from '@/shared/ui/block-component/block-component';
 import { createPagination } from '@/widgets/pagination/ui/pagination';
-import { winnerTable } from '@/widgets/winner-table/winner-table';
+import { winnerTable } from '@/widgets/winner-table/ui/winner-table';
 import { initWinners } from '@/pages/model/init';
 import { getWinnersWithCars } from '@/features/get-winners-with-cars/get-winners-with-cars';
 import type { WinnersWithCars } from '@/pages/model/types';
