+++ src/widgets/winner-table/ui/winner-table.ts
